+++ changelog.md
@@ -1,3 +1,7 @@
+#### 0.6.25 - 11/2/2019 ####
+Geology bonus is now preserved on MAD reset<br>
+Key Mappings are now configurable for multiplier and queue keys<br>
+
 #### 0.6.24 - 10/31/2019 ####
 The Halloween feat will now unlock inside a micro universe<br>